refactor(solver): name constraint walker guard states
