+++ crates/tsz-solver/src/operations/constraints/mod.rs
@@ -9,4 +9,6 @@ mod indexed_access;
 mod reverse_mapped;
 mod signatures;
 mod walker;
+mod walker_guard_state;
 mod walker_helpers;
+mod walker_support;
