+++ crates/tsz-solver/src/operations/constraints/walker.rs
@@ -1,215 +1,15 @@
-//! Core type constraint walker for generic type inference.
-//!
-//! Contains the main structural walker (`constrain_types` / `constrain_types_impl`)
-//! that collects type constraints when inferring generic type parameters from
-//! argument types.
-
-use crate::def::DefId;
 use crate::inference::infer::InferenceContext;
 use crate::instantiation::instantiate::{TypeSubstitution, instantiate_type};
-use crate::operations::core::MAX_CONSTRAINT_STEPS;
-use crate::operations::{AssignabilityChecker, CallEvaluator, MAX_CONSTRAINT_RECURSION_DEPTH};
+use crate::operations::constraints::walker_guard_state::with_placeholder_visited;
+use crate::operations::{AssignabilityChecker, CallEvaluator};
 use crate::types::{
     FunctionShape, IntrinsicKind, LiteralValue, ParamInfo, PropertyInfo, TemplateSpan, TypeData,
-    TypeId, TypeListId, TypeParamInfo, TypePredicate,
+    TypeId, TypeParamInfo, TypePredicate,
 };
 use rustc_hash::{FxHashMap, FxHashSet};
-use std::cell::RefCell;
-use std::sync::Arc;
 use tracing::{debug, trace};
 
-// Reusable scratch `FxHashSet<TypeId>` for the five `type_contains_placeholder`
-// call-sites in this module. Each call previously allocated a fresh set;
-// pooling shaves the allocator round-trip plus 2–4 grows. Mirrors the pool
-// pattern from #4722 / #4790 / #4801 / #4805 / #4807.
-thread_local! {
-    static PLACEHOLDER_VISITED_POOL: RefCell<Option<FxHashSet<TypeId>>> =
-        const { RefCell::new(None) };
-}
-
-#[inline]
-pub(super) fn with_placeholder_visited<R>(f: impl FnOnce(&mut FxHashSet<TypeId>) -> R) -> R {
-    let mut visited = PLACEHOLDER_VISITED_POOL
-        .with(|p| p.borrow_mut().take())
-        .unwrap_or_default();
-    visited.clear();
-    let r = f(&mut visited);
-    PLACEHOLDER_VISITED_POOL.with(|p| {
-        let mut slot = p.borrow_mut();
-        let keep = match &*slot {
-            None => true,
-            Some(existing) => visited.capacity() >= existing.capacity(),
-        };
-        if keep {
-            *slot = Some(visited);
-        }
-    });
-    r
-}
-
 impl<'a, C: AssignabilityChecker> CallEvaluator<'a, C> {
-    /// Structural walker to collect constraints: source <: target
-    pub(crate) fn constrain_types(
-        &mut self,
-        ctx: &mut InferenceContext,
-        var_map: &FxHashMap<TypeId, crate::inference::infer::InferenceVar>,
-        source: TypeId,
-        target: TypeId,
-        priority: crate::types::InferencePriority,
-    ) {
-        {
-            let steps = self.constraint_step_count.get();
-            if steps >= MAX_CONSTRAINT_STEPS {
-                return;
-            }
-            self.constraint_step_count.set(steps + 1);
-        }
-
-        if !self.constraint_pairs.borrow_mut().insert((source, target)) {
-            return;
-        }
-
-        // Check and increment recursion depth to prevent infinite loops
-        {
-            let depth = self.constraint_recursion_depth.get();
-            if depth >= MAX_CONSTRAINT_RECURSION_DEPTH {
-                // Safety limit reached - return to prevent infinite loop
-                return;
-            }
-            self.constraint_recursion_depth.set(depth + 1);
-        }
-
-        // Perform the actual constraint collection
-        self.constrain_types_impl(ctx, var_map, source, target, priority);
-
-        // Decrement depth on return
-        self.constraint_recursion_depth
-            .set(self.constraint_recursion_depth.get() - 1);
-    }
-
-    /// Constrain a mapped type's parameters from a source whose key space is
-    /// empty — the intrinsic `object` type, or an empty object literal `{}`.
-    ///
-    /// `keyof <keyless source>` is `never`, so the mapped key space collapses to
-    /// `never`. For a non-homomorphic mapped type (`{ [P in K]: T }`, whose
-    /// constraint is a plain key set rather than `keyof <param>`) the value space
-    /// is empty as well, so the template type parameter is inferred as `never`.
-    /// This matches tsc's `inferToMappedType`, which infers `[never, never]` for
-    /// both `rec(object)` and `rec({})` against
-    /// `<K extends PropertyKey, T>(r: Record<K, T>)`. Homomorphic mapped types
-    /// (constraint `keyof <param>`) keep only the key-space collapse, preserving
-    /// their reverse-mapped candidate handling.
-    fn constrain_empty_keyspace_mapped(
-        &mut self,
-        ctx: &mut InferenceContext,
-        var_map: &FxHashMap<TypeId, crate::inference::infer::InferenceVar>,
-        mapped: &crate::types::MappedType,
-        priority: crate::types::InferencePriority,
-    ) {
-        self.constrain_types(ctx, var_map, TypeId::NEVER, mapped.constraint, priority);
-        // Only a homomorphic mapped type (`{ [P in keyof T]: … }` over an
-        // inference placeholder) has a reverse-mapped candidate path to
-        // preserve. For any other (non-homomorphic) mapped type the value space
-        // is empty too, so the template type parameter is inferred as `never`.
-        if self
-            .find_keyof_inference_target(mapped.constraint, var_map)
-            .is_none()
-        {
-            let subst = TypeSubstitution::single(mapped.type_param.name, TypeId::NEVER);
-            let instantiated_template = instantiate_type(self.interner, mapped.template, &subst);
-            self.constrain_types(
-                ctx,
-                var_map,
-                TypeId::NEVER,
-                instantiated_template,
-                crate::types::InferencePriority::MappedType,
-            );
-        }
-    }
-
-    /// Normalize a union's members for inference partitioning by peeling
-    /// transparent identity-alias applications (`type Some<X> = X`).
-    ///
-    /// A target arm like `Some<A>` is otherwise an opaque `Application` that
-    /// never registers as the naked inference variable `A`, so the single-naked
-    /// variable union strategy (`inferToMultipleTypes`) cannot route unmatched
-    /// source constituents into `A` and it collapses to its default. Peeling the
-    /// arm to its forwarded type fixes that; a peel that yields a union is
-    /// flattened one level so downstream partitioning sees the leaf members.
-    fn normalize_union_members_for_inference(&mut self, members: TypeListId) -> Arc<[TypeId]> {
-        let original = self.interner.type_list(members);
-        // Cheap pre-scan: only a `Lazy`-alias `Application` can be an identity
-        // alias. When the union has none — the overwhelming common case on this
-        // hot constraint path — return the interned list untouched, avoiding
-        // both a fresh allocation and any resolver work.
-        if !original
-            .iter()
-            .any(|&member| self.lazy_alias_application_def_id(member).is_some())
-        {
-            return original;
-        }
-        let mut out = Vec::with_capacity(original.len());
-        for &member in original.iter() {
-            let Some(peeled) = self.try_peel_identity_alias_application(member) else {
-                out.push(member);
-                continue;
-            };
-            if let Some(TypeData::Union(inner)) = self.interner.lookup(peeled) {
-                out.extend(self.interner.type_list(inner).iter().copied());
-            } else {
-                out.push(peeled);
-            }
-        }
-        out.into()
-    }
-
-    /// The alias `DefId` when `member` is an `Application` whose base is a `Lazy`
-    /// alias reference — the only shape `try_peel_identity_alias_application` can
-    /// peel. Resolver-free, so it doubles as the cheap pre-scan gate.
-    fn lazy_alias_application_def_id(&self, member: TypeId) -> Option<DefId> {
-        let TypeData::Application(app_id) = self.interner.lookup(member)? else {
-            return None;
-        };
-        let base = self.interner.type_application(app_id).base;
-        if base.is_intrinsic() {
-            return None;
-        }
-        match self.interner.lookup(base)? {
-            TypeData::Lazy(def_id) => Some(def_id),
-            _ => None,
-        }
-    }
-
-    /// If `member` applies a *transparent identity alias* — a type alias whose
-    /// body is exactly one of its own type parameters, e.g. `type Some<X> = X` —
-    /// expand it to the forwarded type argument.
-    ///
-    /// Returns `None` for any alias whose body adds structure
-    /// (`type Box<X> = { v: X }`, `type Opt<X> = X | undefined`): those must stay
-    /// opaque so structural union inference can still match against the wrapper.
-    fn try_peel_identity_alias_application(&mut self, member: TypeId) -> Option<TypeId> {
-        let def_id = self.lazy_alias_application_def_id(member)?;
-        // The alias body must be a bare type parameter declared by this very
-        // alias. Scope the resolver borrow so `expand_type_alias_application`
-        // (which needs `&mut self.checker`) can run afterwards.
-        let is_identity = {
-            let resolver = self.checker.type_resolver()?;
-            let type_params = resolver.get_lazy_type_params(def_id)?;
-            let body = resolver.resolve_lazy(def_id, self.interner)?;
-            match self.interner.lookup(body)? {
-                TypeData::TypeParameter(body_param) => {
-                    type_params.iter().any(|p| p.name == body_param.name)
-                }
-                _ => false,
-            }
-        };
-        if !is_identity {
-            return None;
-        }
-        let peeled = self.checker.expand_type_alias_application(member)?;
-        (peeled != member).then_some(peeled)
-    }
-
     /// Inner implementation of `constrain_types`
     pub(super) fn constrain_types_impl(
         &mut self,
@@ -2196,15 +1996,5 @@ impl<'a, C: AssignabilityChecker> CallEvaluator<'a, C> {
             }
             _ => {}
         }
-    }
-
-    fn constraint_is_nullable_union(&self, constraint: TypeId) -> bool {
-        let Some(TypeData::Union(members)) = self.interner.lookup(constraint) else {
-            return false;
-        };
-        self.interner
-            .type_list(members)
-            .iter()
-            .any(|&member| member.is_nullable())
     }
 }
